+++ PesterExplorer/Private/Get-TitlePanel.ps1
@@ -27,16 +27,24 @@ function Get-TitlePanel {
     param(
         $Item
     )
-    # TODO: HeyItsGilbert/PesterExplorer#4 Add More info!
-    $titles = @(
-        "Pester Explorer - [gray]$(Get-Date)[/]"
-    )
+    $rows = @()
+    $title = "Pester Explorer - [gray]$(Get-Date)[/]"
     if($null -ne $Item){
         $objectName = Format-PesterObjectName -Object $Item
         # Print what type it is and it's formatted name.
-        $titles += "$($Item.GetType().Name): $($objectName)"
+        $title += " | $($Item.GetType().Name): $($objectName)"
     }
-    return $titles | Format-SpectreRows |
+    $rows += $title
+    #region Breakdown
+    $data = @()
+    $data += New-SpectreChartItem -Label "Passed" -Value ($Item.PassedCount) -Color "Green"
+    $data += New-SpectreChartItem -Label "Inconclusive" -Value ($Item.InconclusiveCount) -Color "Grey"
+    $data += New-SpectreChartItem -Label "Skipped" -Value ($Item.SkippedCount) -Color "Yellow"
+    $data += New-SpectreChartItem -Label "Failed" -Value ($Item.FailedCount) -Color "Red"
+    $rows += Format-SpectreBreakdownChart -Data $data -ShowPercentage
+    #endregion Breakdown
+
+    return $rows | Format-SpectreRows |
         Format-SpectreAligned -HorizontalAlignment Center -VerticalAlignment Middle |
         Format-SpectrePanel -Expand
 }